requirements-dev.txt
alabaster==0.7.12
appdirs==1.4.4
attrs==21.4.0
Babel==2.10.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
docutils==0.17.1
fake-bpy-module-2.82==20200514
flake8==4.0.1
flake8-black==0.3.2
flake8-bugbear==22.4.25
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.9
pytz==2022.1
regex==2022.4.24
requests==2.27.1
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.3
typing==3.7.4.3
urllib3==1.26.9
zipp==3.8.0
